requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
accelerate=0.6.2=pypi_0
adapter-transformers=3.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
ca-certificates=2022.3.29=h06a4308_0
certifi=2021.10.8=py38h06a4308_2
charset-normalizer=2.0.12=pypi_0
click=8.1.2=pypi_0
datasets=2.0.0=pypi_0
dill=0.3.4=pypi_0
filelock=3.6.0=pypi_0
frozenlist=1.3.0=pypi_0
fsspec=2022.3.0=pypi_0
huggingface-hub=0.5.0=pypi_0
idna=3.3=pypi_0
joblib=1.1.0=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
multidict=6.0.2=pypi_0
multiprocess=0.70.12.2=pypi_0
ncurses=6.3=h7f8727e_2
numpy=1.22.3=pypi_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.4.2=pypi_0
pip=21.2.4=py38h06a4308_0
pyarrow=7.0.0=pypi_0
pyparsing=3.0.7=pypi_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytz=2022.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.3.15=pypi_0
requests=2.27.1=pypi_0
responses=0.18.0=pypi_0
sacremoses=0.0.49=pypi_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.38.2=hc218d9a_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.11.6=pypi_0
torch=1.11.0=pypi_0
tqdm=4.64.0=pypi_0
typing-extensions=4.1.1=pypi_0
urllib3=1.26.9=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zlib=1.2.11=h7f8727e_4
